First working version of a zenodo schema.
The schema has been developed using ideas from
[asdf](http://asdf-standard.readthedocs.io/en/latest/intro.html) format.

It used a YAML file instead of json to create the schema but behind it
it is the json schema which is used.
Using yaml instead of json allowed more flexibility (comments...)

The limitation of the actual schema are linked to the need of some
keyword if another one has a certain value (e.g: if "upload_type" is
"publication", the keyword  "publication_type" should be present).