Add zenodo metadata classes to validate the metadata for the upload
Contains class and method which allows validation and proper return message
if any problem on the metadata used to upload on Zenodo before the upload.
It is needed since Zenodo does not provide a good error message if upload
failed.
This should use the metadata-1.0.0.yml file in schemas/zenodo
as reference schemas but another schema can be provided.