Complete rewrite on the metadata validation for Zenodo data repository.
The validation is using a jsconschema method (through a passage from YAML).
YAML was used as schema langage because it is more human readable and
also allows comments in the file.

The unit test associated are present as well as a convenient script to test
a metadata file with the yaml schema provided (it is similar to jsonschema
(and it is using it) but be careful, error can still be present in the metadata.
There are some validation (license for example) which has to be done in the software.

There are an update of the zenodo schema description to reflect better the documentation.

Finally there are a Markdown file which contains some informations on how the yaml schema
file was construct.