Modify datalight to zip file containing the data if needed (default)
Zenodo does not allowed upload of directory by consequence
the structure of the data cannot be contains. In addition to that
Zenodo does not have a possibility to download in batch the data
on a record (through the web interface). Create a zip file with
the data solve this two problems and since zenodo is showing
a preview of what is inside of an archive it is easy to see what
the archive contains.

Cleaning of the docopt (calling arguments) and better crash handling with
better error message.